pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdftoxl"
version = "0.0.1"
description = "PDF to EAB-Excel pipeline evals harness"
requires-python = ">=3.11"
dependencies = [
    "openpyxl>=3.1",
    "pydantic>=2.5",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rapidfuzz>=3.6",
    "python-dateutil>=2.8",
    "python-dotenv>=1.0",
    "structlog>=24.1",
    "boto3>=1.34",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.5"]

[project.scripts]
pdftoxl-evals = "pdftoxl.evals.cli:app"
pdftoxl = "pdftoxl.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"src/pdftoxl/evals/cli.py" = ["B008"]
"src/pdftoxl/cli.py" = ["B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
pythonpath = ["src"]
